inline rejoin_test, rejoin_test_large_state and state_sync_malicious_chunk_test
