+++ rs/tests/message_routing/Cargo.toml
@@ -23,6 +23,7 @@ ic-types = { path = "../../types/types" }
 ic-utils = { workspace = true }
 rand = { workspace = true }
 rand_chacha = { workspace = true }
+rejoin-test-lib = { path = "./rejoin_test_lib" }
 
 [[bin]]
 name = "global_reboot_test"
@@ -31,3 +32,15 @@ path = "global_reboot_test.rs"
 [[bin]]
 name = "memory_safety_test"
 path = "memory_safety_test.rs"
+
+[[bin]]
+name = "rejoin_test"
+path = "rejoin_test.rs"
+
+[[bin]]
+name = "rejoin_test_large_state"
+path = "rejoin_test_large_state.rs"
+
+[[bin]]
+name = "state_sync_malicious_chunk_test"
+path = "state_sync_malicious_chunk_test.rs"
