+++ rs/tests/message_routing/rejoin_test_lib/Cargo.toml
@@ -1,0 +1,29 @@
+[package]
+name = "rejoin-test-lib"
+version.workspace = true
+authors.workspace = true
+edition.workspace = true
+description.workspace = true
+documentation.workspace = true
+
+[dependencies]
+canister-test = { path = "../../../rust_canisters/canister_test" }
+dfn_candid = { path = "../../../rust_canisters/dfn_candid" }
+ic-system-test-driver = { path = "../../driver" }
+itertools = { workspace = true }
+slog = { workspace = true }
+tokio = { workspace = true }
+xnet-test = { path = "../../../rust_canisters/xnet_test" }
+candid = { workspace = true }
+ic-agent = { workspace = true }
+ic-base-types = { path = "../../../types/base_types" }
+ic-utils = { workspace = true }
+rand = { workspace = true }
+rand_chacha = { workspace = true }
+dfn_json = { path = "../../../rust_canisters/dfn_json" }
+chrono = { workspace = true }
+futures = { workspace = true }
+
+[lib]
+name = "rejoin_test_lib"
+path = "rejoin_test_lib.rs"
